backend/requirements.txt
# Core dependencies
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4

# Data processing
pandas==2.1.3
numpy==1.26.2

# Zoho CRM SDK
zohocrmsdk8_0==2.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0 